data-jpa/src/main/java/me/jamie/datajpa/repository/MemberJpaRepository.java
package me.jamie.datajpa.repository;

import lombok.RequiredArgsConstructor;
import me.jamie.datajpa.domain.Member;
import org.springframework.stereotype.Repository;

import javax.persistence.EntityManager;

@RequiredArgsConstructor
@Repository
public class MemberJpaRepository {

    private final EntityManager em;

    public Member save(Member member) {
        em.persist(member);
        return member;
    }

    public Member find(Long id) {
        return em.find(Member.class, id);
    }
}
